requirements.txt
flask
redis
aiohttp
cassandra-driver